Begin moving WheelCache to pip.cache
